[TW] 보낸쪽지함, 받은쪽지함 추가

+++ choongangGroup/src/main/java/com/oracle/choongangGroup/taewoo/service/MessageService.java
@@ -42,7 +42,7 @@ public class MessageService {
 		return MessageDto.toDto(message);
 	}
 
-	// 보낸 쪽지함
+	// 받은 쪽지함
 	public List<MessageDto> receiveMessage(Member member, Pageable pageable) {
 //		   List<Message> messages = messageRepository.findAllByReceiver(member);	
 		   List<Message> messages = messageRepository.findByReceiverOrderByMessageIdDesc(member,  pageable);
@@ -56,16 +56,53 @@ public class MessageService {
 	        }
 	        return messageDtos;
 	}
+	// 보낸 쪽지함
+	public List<MessageDto> senderMessage(Member member, Pageable pageable) {
+		 List<Message> messages = messageRepository.findBySenderOrderByMessageIdDesc(member, pageable);
+	        List<MessageDto> messageDtos = new ArrayList<>();
 
-	// 쪽지 삭제
-	public int delete(Long messageId) {
+	        for(Message message : messages) {
+	            // message 에서 받은 편지함에서 삭제하지 않았으면 보낼 때 추가해서 보내줌
+	            if(!message.isDeletedBySender()) {
+	                messageDtos.add(MessageDto.toDto(message));
+	            }
+	        }
+	        return messageDtos;
+	}
+
+	// 받은 쪽지 삭제
+	public int delete(Long messageId, Member member) {
+		Message message = messageRepository.findByMessageId(messageId);
 		System.out.println("delete start....");
 		System.out.println(messageId);
-		
-		
-		int result = messageRepository.deleteByMessageId(messageId);
+		System.out.println(message.isDeletedByReceiver());
+		System.out.println(message.isDeleted());
+		message.deletedByReceiver();		
+		if(message.isDeleted()) {
+			int result = messageRepository.deleteByMessageId(messageId);
+			return result;
+			}
+			int result = messageRepository.updateDeleteReceiver(messageId);
+			return result;
+	}
+	
+	// 보낸 쪽지 삭제
+	public int senderDelete(Long messageId, Member member) {
+		Message message = messageRepository.findByMessageId(messageId);
+		System.out.println("senderDelete start....");
+		System.out.println("messageId -->" + messageId);
+		System.out.println("message -->" + message);
+		System.out.println(message.isDeletedBySender());
+		System.out.println(message.isDeleted());
+		message.deleteBySender();
+		if(message.isDeleted()) {
+			int result = messageRepository.deleteByMessageId(messageId);
+			return result;
+		}
+		int result = messageRepository.updateDeleteSender(messageId);
 		return result;
 	}
+
 
 	// 상세 쪽지
 	public Message Detail(Long messageId) {
@@ -74,6 +111,10 @@ public class MessageService {
 		Message message = messageRepository.findById(messageId).orElseThrow(() -> new NoSuchElementException());
 		return message;
 	}
+
+
+
+
